Add boot copy scripts for i.MX9352 board
This commit introduces two new shell scripts, `copy_boot.sh` and `copy.sh`, to facilitate the copying of boot and filesystem images for the i.MX9352 board. The scripts automate the process of transferring the necessary binary files to the designated boot media, enhancing the setup process for the board.

Signed-off-by: [Your Name] <your.email@example.com>
Signed-off-by: hzx <1289514315@qq.com>