Update device tree and configuration for i.MX9352 board
This commit enhances the device tree source (DTS) for the i.MX9352 board by modifying temperature thresholds for CPU alerts and critical conditions, adding watchdog timers, and refining GPIO configurations. It also updates the defconfig to include the ATF package for the i.MX9352. Additionally, the pin function header is introduced to support the new device tree structure, ensuring better compatibility and functionality for the board.

Signed-off-by: [Your Name] <your.email@example.com>
Signed-off-by: hzx <1289514315@qq.com>
Signed-off-by: Guanghai Wu <guanghai.wu@hcrobots.com>